Calendar: Fix the unit test failure.
As there maybe some edit for DateUtils API on the different android version,
and it make the test cases' expected result maybe not same as the actual
result. So we will print the warning in the log for the test result if this
test case used to test the date format.

CRs-fixed: 652947

Change-Id: Ie2aa984191fb8b5b6c7316235672630291717ed4